Added option to use a zip file as "repository" besides a GIT or SVN repo
The current repo options that we had (SVN and GIT) will always get the trunk version of the component (At least until we implement versioning). But what if instead of trunk, I wanted say release 1.0.0 of aws SDK?  Just that version, not a newer one.

Now with this commit you can do:
tms repo-register zipfile https://github.com/landgraf-dev/aws-sdk-delphi/archive/refs/tags/1.0.0.zip
tms install *aws*

And it will install with this particular release: https://github.com/landgraf-dev/aws-sdk-delphi/releases/tag/1.0.0
(and yes, it is already working :))

But this new "zipfile" protocol might have other uses too. Basically, it allows you to get any smartsetup bundle that is hosted in a website. It will just download the zip from that website, then unzip it and build it as a normal product.

You might create a registry with zip files for your organization with specific bundles for each product, and get them from there.

The only restriction is that the webserver has to return [etags](https://developer.mozilla.org/en-US/docs/Web/HTTP/Reference/Headers/ETag) so we can know if there was an update to the component or not (and not just download blindly every time you want to update). GitHub at least returns etags, so any zip/.tar.ztd hosted in github should work.

+++ tms/src/VCS/Git.HTTPSDownload.pas
@@ -7,11 +7,11 @@ type
 
   GitDownloader = class
   private
-    class procedure GetFile(const DownloadUrl, FileNameOnDisk: string; DownloadLogger: TDownloadLogger); static;
+    class procedure GetFile(const DownloadUrl, FileNameOnDisk: string; DownloadLogger: TDownloadLogger; const ForceDownload: boolean); static;
 
   public
     //Downloads a full githubrepo to a file.
-    class procedure GetRepo(const DownloadUrl, FileNameOnDisk: string; DownloadLogger: TDownloadLogger); static;
+    class procedure GetRepo(const DownloadUrl, FileNameOnDisk: string; DownloadLogger: TDownloadLogger; const ForceDownload: boolean = false); static;
 
   end;
 
@@ -22,7 +22,7 @@ uses Classes, SysUtils, System.Net.HttpClient, UTmsBuildSystemUtils,
 { GitDownloader }
 
 //Code is from TParallelDownloader.DownloadFile. We could unify it, but for now I prefer to evolve it separately.
-class procedure GitDownloader.GetFile(const DownloadUrl, FileNameOnDisk: string; DownloadLogger: TDownloadLogger);
+class procedure GitDownloader.GetFile(const DownloadUrl, FileNameOnDisk: string; DownloadLogger: TDownloadLogger; const ForceDownload: boolean);
 begin
   var Client: THTTPClient := THTTPClient.Create;
   try
@@ -33,7 +33,7 @@ begin
     var ETagFileName := FileNameOnDisk + '.etag';
     var ETag := 'invalid';
     try
-      if TFile.Exists(ETagFileName) then ETag := TFile.ReadAllText(ETagFileName, TEncoding.UTF8);
+      if TFile.Exists(ETagFileName) and not ForceDownload then ETag := TFile.ReadAllText(ETagFileName, TEncoding.UTF8);
     except on ex: Exception do //invalid etag file?
       begin
         DownloadLogger(TVerbosity.info, 'Invalid ETag Cache: ' + ex.Message);
@@ -80,9 +80,12 @@ begin
           DeleteFileOrMoveToLocked(Config.Folders.LockedFilesFolder, FileNameOnDisk);
           RenameFile(TempFileName, FileNameOnDisk);
 
-          var NewETag := AResponse.HeaderValue['ETag'];
-          if NewETag = '' then DownloadLogger(TVerbosity.Error, 'Server at url "' + DownloadUrl + '" doesn''t support ETags. USE ONLY SERVERS WITH ETAGs to avoid continually downloading the same file.');
-          TFile.WriteAllText(ETagFileName, NewETag, TEncoding.UTF8);
+          if not ForceDownload then
+          begin
+            var NewETag := AResponse.HeaderValue['ETag'];
+            if NewETag = '' then DownloadLogger(TVerbosity.Error, 'Server at url "' + DownloadUrl + '" doesn''t support ETags. USE ONLY SERVERS WITH ETAGs to avoid continually downloading the same file.');
+            TFile.WriteAllText(ETagFileName, NewETag, TEncoding.UTF8);
+          end;
         end
 
 
@@ -107,9 +110,9 @@ begin
 end;
 
 class procedure GitDownloader.GetRepo(const DownloadUrl,
-  FileNameOnDisk: string; DownloadLogger: TDownloadLogger);
+  FileNameOnDisk: string; DownloadLogger: TDownloadLogger; const ForceDownload: boolean = false);
 begin
-  GetFile(DownloadUrl, FileNameOnDisk, DownloadLogger);
+  GetFile(DownloadUrl, FileNameOnDisk, DownloadLogger, ForceDownload);
 end;
 
 end.
